Added multiplayer support. It works, but fucks up on slow connections. I'm gonna try to redo the system in a more reliable way, but might fail, so I'll commit this to go back to in case I want to.

+++ app/src/main/java/com/MightyBarbet/LineMayhem/GameObject.java
@@ -7,8 +7,8 @@ import android.graphics.Rect;
  */
 public abstract class GameObject {
 
-    protected int x;
-    protected int y;
+    protected short x;
+    protected short y;
     protected double speedX, accelX;
     protected double speedY, accelY;
     protected int width;
@@ -37,11 +37,11 @@ public abstract class GameObject {
 
 
     //Setters
-    public void setX(int x){
+    public void setX(short x){
         this.x = x;
     }
 
-    public void setY(int y){
+    public void setY(short y){
         this.y = y;
     }
